deliberate failure, test windows build
